feat: add animation for switch swap button
