+++ widget/embedded/src/components/SwitchFromAndTo/index.ts
@@ -1,0 +1,1 @@
+export { SwitchFromAndToButton } from './SwitchFromAndTo';
